pyproject.toml
[project]
name = "airtbench-code"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiodocker>=0.24.0",
    "cyclopts>=4.2.0",
    "docker>=7.1.0",
    "dreadnode==2.0.11",
    "ipykernel>=6.29.5",
    "litellm>=1.83.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "pythonnet>=3.0.5",
    "tenacity>=9.1.2",
    "universal-pathlib>=0.2.6",
]

[project.scripts]
airtbench = "airtbench.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["airtbench"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.9",
    "types-PyYAML>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
extend-exclude = [
    ".hooks",
    "airtbench/challenges/*.ipynb",
    "notebooks/*.ipynb",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "E501",   # line too long (we make best effort)
    "TRY003", # long messages in exception classes
    "EM",     # Picky message construction for exceptions
    "C90",    # mccabe complexity
    "A002",   # shadowing built-in
    "ANN401", # Any type
    "ASYNC109", # timeout arguments are part of the kernel API
    "BLE001", # existing harness cleanup logs broad failures and continues
    "D",      # docstrings
    "PLC0415", # local imports avoid import cycles in cleanup paths
    "PLR0911", # legacy harness control flow
    "PLR0912", # legacy harness control flow
    "PLR0915", # legacy harness control flow
    "PLR2004", # status codes and cleanup intervals are clearer inline here
    "PYI034", # async context manager returns initialized self
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
skip-magic-trailing-comma = false

[tool.uv]
exclude-newer = "2026-04-10T00:00:00Z"
